#119: Move endpoints dict to a separate file
- This change is to prevent a circular import when get_entity_object_from_name is modified
- Also removes a TODO comment that has previously addressed - it was regarding camelCasing of modId etc.
