+++ datagateway_api/src/resources/entities/entity_endpoint_dict.py
@@ -1,0 +1,41 @@
+# endpoint_name: entity_name
+endpoints = {
+    "Applications": "Application",
+    "DataCollectionDatafiles": "DataCollectionDatafile",
+    "DataCollectionDatasets": "DataCollectionDataset",
+    "DataCollectionParameters": "DataCollectionParameter",
+    "DataCollections": "DataCollection",
+    "DatafileFormats": "DatafileFormat",
+    "DatafileParameters": "DatafileParameter",
+    "Datafiles": "Datafile",
+    "DatasetParameters": "DatasetParameter",
+    "DatasetTypes": "DatasetType",
+    "Datasets": "Dataset",
+    "Facilities": "Facility",
+    "FacilityCycles": "FacilityCycle",
+    "Groupings": "Grouping",
+    "InstrumentScientists": "InstrumentScientist",
+    "Instruments": "Instrument",
+    "InvestigationGroups": "InvestigationGroup",
+    "InvestigationInstruments": "InvestigationInstrument",
+    "InvestigationParameters": "InvestigationParameter",
+    "InvestigationTypes": "InvestigationType",
+    "InvestigationUsers": "InvestigationUser",
+    "Investigations": "Investigation",
+    "Jobs": "Job",
+    "Keywords": "Keyword",
+    "ParameterTypes": "ParameterType",
+    "PermissibleStringValues": "PermissibleStringValue",
+    "PublicSteps": "PublicStep",
+    "Publications": "Publication",
+    "RelatedDatafiles": "RelatedDatafile",
+    "Rules": "Rule",
+    "SampleParameters": "SampleParameter",
+    "SampleTypes": "SampleType",
+    "Samples": "Sample",
+    "Shifts": "Shift",
+    "Studies": "Study",
+    "StudyInvestigations": "StudyInvestigation",
+    "UserGroups": "UserGroup",
+    "Users": "User",
+}
