+++ datagateway_api/src/resources/entities/entity_map.py
@@ -2,50 +2,7 @@ import datetime
 
 from sqlalchemy.inspection import inspect
 
-from datagateway_api.common.database.models import EntityHelper
-
-
-# endpoint_name: entity_name
-endpoints = {
-    "Applications": "Application",
-    "DataCollectionDatafiles": "DataCollectionDatafile",
-    "DataCollectionDatasets": "DataCollectionDataset",
-    "DataCollectionParameters": "DataCollectionParameter",
-    "DataCollections": "DataCollection",
-    "DatafileFormats": "DatafileFormat",
-    "DatafileParameters": "DatafileParameter",
-    "Datafiles": "Datafile",
-    "DatasetParameters": "DatasetParameter",
-    "DatasetTypes": "DatasetType",
-    "Datasets": "Dataset",
-    "Facilities": "Facility",
-    "FacilityCycles": "FacilityCycle",
-    "Groupings": "Grouping",
-    "InstrumentScientists": "InstrumentScientist",
-    "Instruments": "Instrument",
-    "InvestigationGroups": "InvestigationGroup",
-    "InvestigationInstruments": "InvestigationInstrument",
-    "InvestigationParameters": "InvestigationParameter",
-    "InvestigationTypes": "InvestigationType",
-    "InvestigationUsers": "InvestigationUser",
-    "Investigations": "Investigation",
-    "Jobs": "Job",
-    "Keywords": "Keyword",
-    "ParameterTypes": "ParameterType",
-    "PermissibleStringValues": "PermissibleStringValue",
-    "PublicSteps": "PublicStep",
-    "Publications": "Publication",
-    "RelatedDatafiles": "RelatedDatafile",
-    "Rules": "Rule",
-    "SampleParameters": "SampleParameter",
-    "SampleTypes": "SampleType",
-    "Samples": "Sample",
-    "Shifts": "Shift",
-    "Studies": "Study",
-    "StudyInvestigations": "StudyInvestigation",
-    "UserGroups": "UserGroup",
-    "Users": "User",
-}
+from datagateway_api.src.resources.entities.entity_endpoint_dict import endpoints
 
 
 def type_conversion(python_type):
